+++ projects/sage-angular/src/lib/footer/footer.component.html
@@ -12,7 +12,7 @@
     <div class="sage-footer-version">
       <img class="sage-angular-logo" src="images/angular-white-transparent.svg"
         alt="Sage Angular icon">
-      <span class="version">Current Version: x.y.z</span>
+      <span class="version">Current Version: {{version}}</span>
     </div>
 
     <div class="sage-footer-copyright">
